[FI-561] 회차 생성시 회차 좌석 자동 생성 및 회차 좌석 생성 시 좌석 불러오기 merge

+++ FLOWIN/performance-service/src/main/java/io/why503/performanceservice/domain/showseat/model/dto/request/SeatPolicyRequest.java
@@ -6,5 +6,5 @@ import jakarta.validation.constraints.NotNull;
 public record SeatPolicyRequest (
         @NotBlank String seatArea,  //좌석 구역(예: VIP, A, B)
         @NotBlank String grade,     //공연 좌석 등급
-        @NotNull int price          //공연 좌석 가격
-){ }
+        @NotNull Long price          //공연 좌석 가격
+){}
